refactoring : 파일 그룹화 및 enum 데이터 위치 수정
