+++ lib/presentation/on_boarding/on_boarding_birth/on_boarding_birth_screen.dart
@@ -4,8 +4,8 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:frontend/config/theme/color_data.dart';
 import 'package:frontend/config/theme/text_data.dart';
 import 'package:frontend/presentation/on_boarding/components/black_points.dart';
-import 'package:frontend/presentation/on_boarding/on_boarding_birth_viewmodel.dart';
-import 'package:frontend/presentation/on_boarding/on_boarding_job_screen.dart';
+import 'package:frontend/presentation/on_boarding/on_boarding_birth/on_boarding_birth_viewmodel.dart';
+import 'package:frontend/presentation/on_boarding/on_boarding_job/on_boarding_job_screen.dart';
 import 'package:get/get.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:form_builder_validators/form_builder_validators.dart';
