+++ lib/presentation/on_boarding/on_boarding_nickname/on_boarding_nickname_viewmodel.dart
