+++ lib/res/constants.dart
@@ -1,0 +1,7 @@
+enum Job {
+  student,
+  jobSeeker,
+  officeWorkers,
+  freelancer,
+  etc,
+}
